packages/sdk/pyproject.toml
[project]
name = "rag-debugger"
version = "1.1.1"
description = "Real-time debugging SDK for RAG pipelines"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Chandu Bobbili"},
]
keywords = ["rag", "debugging", "llm", "observability", "tracing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Debuggers",
]
dependencies = [
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
langchain = ["langchain-core>=0.1.0"]
llamaindex = ["llama-index-core>=0.10.0"]
openai = ["openai>=1.0.0"]
all = ["langchain-core>=0.1.0", "llama-index-core>=0.10.0", "openai>=1.0.0"]

[project.urls]
Homepage = "https://github.com/ChanduBobbili/rag-debugger"
Repository = "https://github.com/ChanduBobbili/rag-debugger"
Issues = "https://github.com/ChanduBobbili/rag-debugger/issues"

[tool.hatch.build.targets.wheel]
packages = ["rag_debugger"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
